Refresh intermediate result before finish
It takes a little extra time to finish than the one without intermediate
result.

Current output structure design is a very rough version, will polish it
in the near future.
It can't be compiled on Mac for this version, will fix it later.

refresh tiles from the center

To be finished this evening.