Issue #17 fix. Need some time to be 100% sure.
use_fake_user needs to be set so that Blender will not ignore it if
there is no users when saving.
Failing to store it will cause missing nodes in node_groups which is why
material keeps missing.

So far this is a bug causing the missing material bug, I will need more
time to check if there are others bugs causing the same issue.

Will close the bug if there is no such issue in the next couple of
weeks.